semantic router/requirements.txt
fastapi
uvicorn
pipecat-ai-flows==0.0.3
groq
semantic-router
qdrant-client
langfuse
pydantic
pydantic-settings
python-dotenv
